Improve configuration loading
