+++ packages/config/package.json
@@ -21,7 +21,6 @@
     "find-up": "^4.1.0",
     "map-obj": "^4.1.0",
     "minimist": "^1.2.0",
-    "p-locate": "^4.1.0",
     "path-exists": "^4.0.0",
     "resolve": "^1.12.0"
   },
