+++ packages/config/path.js
@@ -1,12 +1,30 @@
-const findUp = require('find-up')
-const pLocate = require('p-locate')
+const { resolve } = require('path')
+const { promisify } = require('util')
 
-// Retrieve path to the configuration file.
-// Lookup from `cwd` (default: current directory) to find any file named
-// `netlify.toml`, `netlify.yml`, etc.
-const getConfigPath = async function(cwds) {
-  const cwdsA = Array.isArray(cwds) ? cwds : [cwds]
-  const cwd = await pLocate(cwdsA, async cwd => Boolean(await findUp(FILENAMES, { cwd })))
+const findUp = require('find-up')
+const resolvePath = require('resolve')
+const pathExists = require('path-exists')
+
+const pResolve = promisify(resolvePath)
+
+// Configuration location can be:
+//  - `undefined`, in which case `cwd` (default: current directory) is looked up
+//    to find any file named `netlify.toml`, `netlify.yml`, etc.
+//  - a local path
+//  - a Node module. This allows configuration sharing
+const getConfigPath = async function(configFile, cwd) {
+  if (configFile === undefined) {
+    return getDefaultConfig(cwd)
+  }
+
+  if (isNodeModule(configFile)) {
+    return getModuleConfig(configFile, cwd)
+  }
+
+  return getLocalConfig(configFile, cwd)
+}
+
+const getDefaultConfig = async function(cwd) {
   const configPath = await findUp(FILENAMES, { cwd })
 
   if (configPath === undefined) {
@@ -20,4 +38,29 @@ const getConfigPath = async function(cwds) {
 
 const FILENAMES = ['netlify.toml', 'netlify.yml', 'netlify.yaml', 'netlify.json', 'netlify.js']
 
-module.exports = getConfigPath
+const isNodeModule = function(configFile) {
+  return configFile.startsWith('@') || !configFile.includes('/')
+}
+
+// We use `resolve` because `require()` should be relative to `baseDir` not to
+// this `__filename`
+const getModuleConfig = async function(configFile, cwd) {
+  try {
+    return await pResolve(configFile, { basedir: cwd })
+  } catch (error) {
+    error.message = `Configuration file does not exist: ${configFile}\n${error.message}`
+    throw error
+  }
+}
+
+const getLocalConfig = async function(configFile, cwd = process.cwd()) {
+  const configPath = resolve(cwd, configFile)
+
+  if (!(await pathExists(configPath))) {
+    throw new Error(`Configuration file does not exist: ${configPath}`)
+  }
+
+  return configPath
+}
+
+module.exports = { getConfigPath }
